Rename/move 'wormhole' module..
Otherwise this leads to missing import errors when using py2app --
presumably due to namespace-related conflicts with the 'wormhole'
package/dependency
